Update Basic_Data_Types.md
Changes:

Added "for" before "novice programmers"
"on this question" to "about this question"
"on every part" to "at every part"
"steps" to "process"
"what kind of data we are dealing with" to "what kind of data are we dealing with"
Added quotation marks around the final question to maintain consistency.
Changed P to 'P' for consistency in formatting.
Changed Problem statement to Problem Statement for proper noun capitalization.
Added a comma after "output".
Changed "is" to are to agree with the plural subject "kinds".
Changed "as" to a.
Changed Hello World! to 'Hello, World!' for correct punctuation and consistency with single quotation marks.
Changed "into" to on.
Added single quotation marks around string to maintain consistency.
Changed "language" to languages for grammatical agreement.
Added a comma after "data type".
Changed "about our chatbot" to our chatbot for grammatical clarity.
Changed "will be a string data" to will be a string of data for clarity.
Changed "if you create" to you create for consistency.
Changed the comma after "fitness activity" to a semicolon for proper punctuation.
Changed "ask on the number of steps" to ask for the number of steps for grammatical accuracy.
Changed "may not so obviously" to may not seem so obviously.
Added to be after "obviously".
Changed "new line" to newline.
Added characters after "tab" for clarity.
Changed "can be actually" to actually.
Added a before string data type.
Changed "instead of numbers" to instead of a number.
Changed "student id data" to student ID data.
Changed "whole number" to whole numbers.
Changed "decimal numbers represented as floating point" to decimal numbers represented as floating-point numbers.
Changed "example for int data type" to example of the int data type.
Added a before "decimal point".
Changed "In computer" to In computers.
option to change from "1.749999..."to "1.7500000..."
Added a period after "etc" to etc..
 Added "a" before `string` data type.
 Added "a" before `string` data type in the second sentence.
 Changed "*single* double quotes" to "double" quotes for clarity.
 Changed "codes" to "code" for grammatical correctness.. Added "in" before "the above" for clarity.
 Changed "the above that once you press" to "the above example that once you press" for clarity.
 Changed "of our keyboard" to "on your keyboard" for correctness.
 Changed "returns back" to "return" for clarity.
Changed "*single* single quotes" to **single** quotes.
changed "*tripple* double quotes" to **triple** double quotes.
Changed "*tripple* single quotes" to **triple** single quotes.
"quotes" changed to "quote" for singular agreement.
"finds a closing quotes" corrected to "finds a closing quote" for singular agreement.
"Python interpret" changed to "Python interprets" for correct verb form.
"Notice that Python does not interpret World" changed to "Notice that Python does not interpret 'World'" to clarify the reference to the word.
"Therefore, Python try to interpret" changed to "Therefore, Python tries to interpret" for correct verb form.
"But how about the triple quotes?" changed to "But how about the triple quotes?" for correct punctuation.
"What is it used for?" changed to "What are they used for?" for agreement with the plural "triple quotes".
Changed "For example," to "For example," for consistency with the previous formatting.